@@ -3,12 +3,12 @@ source 'http://rubygems.org'
 gemspec
 
 group :test do
-  gem 'lolita', '~> 4.0'
-  gem 'rspec-rails', '~> 2.14'
-  gem 'ffaker', '~> 1.16.2'
-  gem 'factory_girl', '~> 3.6.2'
-  gem 'sqlite3', '~> 1.3.7'
-  gem 'rspec', '~> 2.14'
-  gem 'rails', '~> 3.2'
+  gem 'lolita', git: 'https://github.com/ithouse/lolita', ref: '7f88ce59555940aaa46dd9fd8837a32891fc0280'
+  gem 'rspec-rails', '~> 3.7'
+  gem 'ffaker', '~> 1.0'
+  gem 'factory_bot', '~> 5.2'
+  gem 'sqlite3'
+  gem 'rspec'
+  gem 'rails', '~> 5.2'
   gem 'pry-byebug'
 end
