@@ -1,4 +1,4 @@
-class CreateFiles < ActiveRecord::Migration
+class CreateFiles < ActiveRecord::Migration[4.2]
   class << self
     
     def up
