@@ -9,7 +9,7 @@ end
 require File.expand_path('test_orm/support')
 
 require 'ffaker'
-require 'factory_girl'
+require 'factory_bot'
 require File.expand_path('lib/lolita-file-upload')
 
 require "carrierwave"
@@ -18,6 +18,13 @@ require 'carrierwave/orm/'+current_orm.to_s.gsub("_","")
 
 require File.expand_path("test_orm/#{current_orm}")
 
+RSpec.configure do |config|
+  config.include FactoryBot::Syntax::Methods
+
+  config.before(:suite) do
+    FactoryBot.find_definitions
+  end
+end
 
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
 
